pyproject.toml

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "lava-dl"
packages = [
    {include = "lava", from = "src"},
    {include = "tests"}
]
include = ["tutorials"]
version = "0.4.0.dev0"
readme = "README.md"
description = "A library of deep learning tools, which consists of lava.lib.dl.slayer and lava.lib.dl.netx for training and deployment of event-based deep neural networks on traditional as well as neuromorphic backends. Lava-DL is part of Lava Framework"
homepage = "https://lava-nc.org/"
repository = "https://github.com/lava-nc/lava-dl"
authors = [ 
    "Intel's Neuromorphic Computing Lab and the open source community <lava@intel.com>"
]
license = "(BSD-3-Clause)"
keywords = [
    "neuromorphic",
    "ai",
    "artificial intelligence",
    "neural models",
    "spiking neural networks",
    "deep learning"
]
classifiers = [
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent"
]

[tool.poetry.urls]
"Issue and Bug Tracker" = "https://github.com/lava-nc/lava-dl/issues"
"Questions and Answers" = "https://github.com/lava-nc/lava-dl/discussions/categories/q-a"
"Frequently Asked Questions" = "https://github.com/lava-nc/lava-dl/wiki/Frequently-Asked-Questions-(FAQ)"
"Discussions" = "https://github.com/lava-nc/lava-dl/discussions"

[tool.poetry.dependencies]
python = ">=3.8, <3.11"

lava-nc = { git = "https://github.com/lava-nc/lava.git", branch = "main", develop = true }

torchvision = "^0.15.1"
h5py = "^3.7.0"
ninja = "^1.10.2.3"
matplotlib = "^3.5.2"
numpy = "^1.24.4"
scipy = "^1.8.1"
pillow = ">=10.0.1,<11.0"
pytest = "^7.2.0"
unittest2 = "^1.1.0"
torch = ">=2.0.0, !=2.0.1"
opencv-python-headless = "^4.8.1.78"

[tool.poetry.dev-dependencies]
bandit = "1.7.4"
coverage = "^6.3.2"
darglint = "^1.8.1"
nbconvert = ">=7.2.10, <7.3"
flake8 = "^4.0.1"
flake8-bandit = "^3.0.0"
flake8-bugbear = "^22.1.11"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.8.0"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.2.0"
flake8-isort = "^4.1.1"
flake8-mutable = "^1.2.0"
flake8-pytest-style = "^1.6.0"
flake8-spellcheck = "^0.25.0"
flakeheaven = "^3.2.1"
pep8-naming = "^0.11.1"
poetry = "^1.1.13"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
unittest2 = "^1.1.0"
black = "^22.1.0"

[tool.black]
line-length = 80
target-version = [
    "py38"
]
include = "\\.pyi?$"
exclude = """
(
  /(
      \\.eggs
    | \\.git
    | \\.venv
    | _build
    | build
    | dist
  )/
)
"""

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--strict-markers --cov=src/lava"
testpaths = [
    "tests"
]

[tool.coverage.run]
relative_files = true
source = [
    "src/lava",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplemented()",
    "if __name__ == .__main__.:",
    "main()",
    "parser\\..",
    "argparse\\..",
]
fail_under = 45
show_missing = true

[tool.flakeheaven]
extended_default_ignore=[]  # Fix for bug while using newer flake8 ver.
format = "grouped"
max_line_length = 80
show_source = true
exclude = ["./docs/"]

[tool.flakeheaven.plugins]
flake8-bandit = ["+*", "-S322", "-B101"]       # Enable a plugin, disable specific checks

# Disable below plugins temporarily until lint fix is pushed
flake8-bugbear = ["-*"]
flake8-builtins = ["-*"]
flake8-comprehensions = ["-*"]
flake8-darglint = ["-*"]
flake8-docstrings = ["-*"]
flake8-eradicate = ["-*"]
flake8-isort = ["-*"]
flake8-mutable = ["-*"]
flake8-pytest-style = ["-*"]
flake8-spellcheck = ["-*"]
mccabe = ["-*"]
pep8-naming = ["-*"]

# Excluding until lint fix is pushed
# flake8-bugbear = ["+*"]                     # Enable a plugin
# flake8-builtins = ["+*"]
# flake8-comprehensions = ["+*"]
# flake8-darglint = ["+*"]
# flake8-docstrings = ["+*"]
# flake8-eradicate = ["+*"]
# flake8-isort = ["+*"]
# flake8-mutable = ["+*"]
# flake8-pytest-style = ["+*"]
# flake8-spellcheck = ["-*"]
# mccabe = ["+*"]
# pep8-naming = ["+*"]
# pyflakes = ["+*"]
# pylint = ["+*"]

pycodestyle = ["+*", "-W503", "-E203"]
pyflakes = ["-*"]                           # Disable temporarily until lint fix is pushed
pylint = ["-*"]                             # Disable temporarily until lint fix is pushed

[tool.flakeheaven.exceptions."tests/"]
pycodestyle = ["-F401"]                     # Disable a check
flake8-bandit = ["-S101"]                   # Ignore asserts for tests
pyflakes = ["-*"]                           # Disable a plugin

[tool.flakeheaven.exceptions."src/lava/lib/dl/slayer/object_detection/models/*"]
pycodestyle = ["-E501", "-E241", "-E221"]   # Disable a check

[tool.flakeheaven.exceptions."tutorials/lava/lib/dl/slayer/tiny_yolo_sdnn/train_sdnn.py"]
pycodestyle = ["-E501"]                     # Disable a check

